Fixed PreviewWebGl inheritance, aligned cube size
