+++ src/services/TileService.ts
@@ -5,7 +5,7 @@ import {Subject} from "rxjs";
 
 import {Point2D, TileCoordinate} from "models";
 import {BackendService, TileWebGLService} from "services";
-import {AppStore} from "stores";
+import {AppStore, PREVIEW_PV_FILEID} from "stores";
 import {copyToFP32Texture, createFP32Texture, GL2} from "utilities";
 
 import ZFPWorker from "!worker-loader!zfp_wrapper";
@@ -89,7 +89,27 @@ export class TileService {
         }
     }
 
-    public decompressPreviewRasterData(compressedView: Uint8Array, eventArgs, previewData: CARTA.PvPreviewData, nanEncodings32: Int32Array) {
+    public decompressPreviewRasterData(previewData: CARTA.PvPreviewData) {
+        const compressedArray = previewData.imageData;
+        const nanEncodings32 = new Int32Array(previewData.nanEncodings.slice(0).buffer);
+        let compressedView = new Uint8Array(Math.max(compressedArray.byteLength, previewData.width * previewData.height * 4));
+        compressedView.set(compressedArray);
+
+        const eventArgs = {
+            fileId: PREVIEW_PV_FILEID,
+            channel: 0,
+            stokes: 0,
+            width: previewData.width,
+            subsetHeight: previewData.height,
+            subsetLength: compressedArray.byteLength,
+            compression: previewData.compressionQuality,
+            nanEncodings: nanEncodings32,
+            tileCoordinate: 0,
+            layer: 0,
+            requestId: 0,
+            previewId: previewData.previewId
+        };
+
         this.workers[0].postMessage(["preview decompress", compressedView.buffer, eventArgs, previewData], [compressedView.buffer, nanEncodings32.buffer]);
     }
